requirements.txt
# requirements.txt
python-telegram-bot
psycopg2-binary
python-dotenv
pyzbar
Pillow
qrcode